refactor: Restructure repository with core/src and plugins/*/src layout
Reorganizes the codebase for monorepo package-ready structure where core
and plugins can be published as separate npm packages.

Structure changes:
- Move core engine from src/ to core/src/ directory
- Move core build configs to core/ directory:
  - core/package.json (core package config)
  - core/tsconfig.json (core TypeScript config)
  - core/tsup.config.ts (core build config)
  - core/jest.config.js (core test config)
- Move plugins from examples/ to individual plugin directories:
  - plugins/physics/src/PhysicsPlugin.ts
  - plugins/spatial-partition/src/SpatialPartitionPlugin.ts
  - plugins/debug-visualizer/src/DebugVisualizerPlugin.ts
  - plugins/profiling/src/ProfilingPlugin.ts
  - plugins/resource-manager/src/ResourceManagerPlugin.ts
- Create root package.json for workspace coordination
- Retain examples/games/ and examples/integrations/ for sample code

Workspace setup:
- Root package.json configured with npm workspaces
- Scripts delegate to core workspace
- Repository-level tooling (husky, lint-staged) at root
- Benchmarks remain at root level

Updated configurations:
- Update all import paths in plugins, examples, and benchmarks
- Update core configs to use relative paths (src/ instead of core/src/)
- Update .gitignore to ignore **/dist (dist in any directory)
- Update CLAUDE.md with new monorepo structure documentation

Benefits:
- Each plugin directory can have its own package.json
- Ready for monorepo tooling (npm workspaces, Lerna, Turborepo, etc.)
- Clear separation between core, plugins, and examples
- Follows standard npm package convention with src/ directories
- Easy to add plugin-specific tests, configs, and documentation
- Core can be published independently

All tests pass (192/192) and build successful.
