+++ .gitignore
@@ -197,4 +197,4 @@ FakesAssemblies/
 /node_modules/
 /.ntvs_analysis.dat
 /coverage.html
-/dist
+**/dist
