+++ core/jest.config.js
