+++ core/package.json
@@ -1,0 +1,60 @@
+{
+  "name": "orion-ecs",
+  "version": "0.1.1-ts",
+  "description": "A lightweight and flexible Entity Component System (ECS) framework written in TypeScript",
+  "main": "dist/orion.js",
+  "types": "dist/orion.d.ts",
+  "scripts": {
+    "build": "tsup",
+    "test": "jest",
+    "lint": "npm run lint:oxlint && npm run lint:biome",
+    "lint:oxlint": "oxlint src",
+    "lint:biome": "biome lint src",
+    "format": "biome format --write .",
+    "format:check": "biome format .",
+    "typecheck": "tsc --noEmit",
+    "check": "npm run typecheck && npm run lint && npm run format:check"
+  },
+  "keywords": [
+    "ecs",
+    "entity-component-system",
+    "game-development"
+  ],
+  "author": {
+    "name": "Tyler Coles",
+    "email": "tyevco@gmail.com"
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/tyevco/orionecs"
+  },
+  "bugs": {
+    "url": "https://github.com/tyevco/orionecs/issues"
+  },
+  "homepage": "https://github.com/tyevco/orionecs#readme",
+  "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
+    "@biomejs/biome": "^2.3.7",
+    "@commitlint/cli": "^20.1.0",
+    "@commitlint/config-conventional": "^20.0.0",
+    "@types/benchmark": "^2.1.5",
+    "@types/jest": "^30.0.0",
+    "babel-jest": "^30.2.0",
+    "benchmark": "^2.1.4",
+    "husky": "^9.1.7",
+    "jest": "30.2.0",
+    "jest-bench": "^29.7.1",
+    "lint-staged": "^16.2.7",
+    "oxlint": "^1.29.0",
+    "seedrandom": "^3.0.5",
+    "ts-jest": "^29.4.5",
+    "tsup": "^8.5.1",
+    "typescript": "^5.9.3"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
