+++ core/src/definitions.ts
