+++ core/src/managers.ts
