+++ core/tsup.config.ts
