+++ examples/games/platformer.ts
@@ -11,8 +11,8 @@
  * - Camera following player
  */
 
-import { EngineBuilder } from '../../src/engine';
-import type { EntityDef } from '../../src/definitions';
+import { EngineBuilder } from '../../core/src/engine';
+import type { EntityDef } from '../../core/src/definitions';
 
 // ============================================================================
 // Components
